+++ front/src/components/membresias/heroMembresias.tsx
@@ -1,0 +1,23 @@
+
+export default function HeroMembresias() {
+return (
+    <div className=" bg-gray-50">
+     
+
+      <div className="container px-4 py-12 mx-auto">
+<div className="mb-12 text-center">
+          <h1 className="mb-4 text-4xl font-bold text-gray-900 font-Title md:text-5xl">
+            Elige tu{" "}
+            <span className="text-transparent bg-gradient-to-r from-orange-500 to-blue-600 bg-clip-text">
+              Plan Profesional
+            </span>
+          </h1>
+          <p className="max-w-2xl mx-auto mb-8 text-xl text-gray-600 font-Text">
+            Selecciona el plan que mejor se adapte a tus necesidades y comienza a hacer crecer tu negocio
+          </p>
+        </div>
+      </div>
+      </div>
+      
+)
+}
